refactor : 사용자 입력에 관한 함수들 모듈화
사용자 입력을 처리하는 함수들을 userInputHandler 파일로 모듈화
